pyproject.toml
[tool.poetry]
name = "Catalog_Dj"
version = "0.1.0"
description = ""
authors = ["Paul <p.bitiakov.official@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
black = "^24.8.0"
pillow = "^11.0.0"
ipython = "^8.29.0"
flake8 = "^7.1.1"
pytils = "^0.4.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
python-dotenv = "^1.0.1"
django = "4.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
